[config] Use static SECRET_KEY from environ.
CSRF token is generated from the secret key.

In the old way, the secret key is randomly generated. Because of this, when the server was restarted, the secret key changed => the CSRF token changed => POST request would fail. This is not good especially for save_answer POST because the return code is 400, not 403 and the exam page would not remind user to login again.