[exam.js] parse exam.finish_date using moment()
exam.finish_date (in GMT+7) used to be parsed using moment.utc(). The parsed result then is passed to the countdown object (let's call it finish_date).

The countdown object calculate the seconds remaining by doing finish_date-moment(). The bug here is that finish_date is in UTC (from GMT+7 by subtracting 420 minutes) and moment() is in
whatever user's timezone is. And it happens when the user's timezone != the original timezone (GMT+7)