Revert "Change domain to thi.tuankiet65.moe"
Turns out I was wrong. I had the impression that in order to get a
certificate for *.example.com I must also be able to get a certificate
for its parent domain (which is example.com). In reality to
get a certificate for thi.quandoansontra.org.vn all I had to do is to
point thi.quandoansontra.org.vn to my server (which I've already done),
and then select the http-01 challenge. Voila, certificate obtained!

This reverts commit 67b343d9b97a198bfb85fb1014387b473e5fc141.