exam.js: Get the current timestamp from the server
During testing on a MacBook today the countdown was always off for a
fixed amount of seconds. Turns out the MacBook's clock was off for
that fixed amount of seconds, and exam.js relies on that incorrect local
time to figure out the remaining time. To fix this, exam_page.html now
includes a server timestamp, and treat this as the current time. Of
course this timestamp might be off for a few seconds due to transmission
delays and such, but since we allow a one-minute grace period to close
the exam, this won't hurt much (unless when it does hurt).